Fixup a type irregularity.
So, we hash ireps. They have their types hashed in. That means
identical expressions with different types will not be found in a hash
map. This is bad, because when we take the address of certain symbols,
they'll get assigned different numbers in the pointer logic object, but
then will just have the symbol name represented in smt.

That amounts to double assignment, and thus always-unsat formula.

So, fix this by during migration, normalising all types to the one that
appears in the symbol table for a symbol. This eliminates cases where we
have a function prototype, take the address of it, then have the address
of the same function after a more fleshed out or different-ish
prototype.

There might be other cases where we develop differing types for symbols,
but I haven't come across them yet.