Refine where terminating execeptions halt.
That is: where we make the guard false. Because right now if we have an
unhandled assertion then we encode the call to the default handler, then
set the guard to false, which makes nothing further execute.

So instead, remove that, and put an assume(false) in the default handler
instead. Thus the same thing is achieved.