Attempt to fix polymorphic read dereferences.
When we have a polymorphic scenario, dereferencing something ends up
with subclasses getting casted up to the superclass, discarding data
along the way. That's all fine and well, but breaks when it hits the
solver, because the struct being casted doesn't have the same format as
the other.

So intead, follow what (essentially) already exists in master, we allow
casts up from subclass to superclass, and only copy across fields that
are common to both of the structs.