Massage type scenario of deref'ed exprs.
I'm experimenting with making names stored in symbol2t's, which requires
that every type for a symbol being consistent. I.E., we can't have a
scenario where we access a symbol (say 'p') with two differing types,
because they'll end up having different hash codes, different name
entries, and thus get renamed to different names.

This can occur when the type in a dereference gets expanded: variables
in the guard to exprs get the expanded pointer type of a variable,
rather than the short "pointer with symbol subtype" form. So, munge it
to only construct it from the originally dereferenced expression.