Start working towards pool allocated ireps.
So the situation is this: we spend years at a time in malloc and free.
By all metrics, it's perhaps 25% of the time. This is almost entirely
due to the detach-based way of creating expressions and duplicating
data.

It'd be better if we used pools: they can allocate a large number of
exprs at a time and just return them as necessary, and if one is freed
it'll be allocated again later without having had to return to the
global heap. Everyone wins. (More memory is allocated in total, but
probably not a fatal amount, and except for multithreading ESBMC never
actually frees exprs en mass).

Trouble: it has to be a type specific pool, because different types are
different sizes. That's fine. However, the dispose method of
shared_ptr's just deletes the pointer with 'delete px', which calls
expr2t's deleter, which can't be type specific. If we were to override
expr2t's deleter, we still wouldn't be able to do anything.

So, we have to start using shared_ptr's deleter mechanism, which is what
I'm adding in here. The intention is that you'll only ever be able to
construct an expression as a result of a something2tc class, and that'll
pass in the appropriate deleter. The non-deleter constructors for
expr2tc are about to disappear.