Massacre the something2tc type scenario, again.
So, we need to make this even more type generic than normal, as we want
to be using this template with both expr2t and type2t situations (and
probably code2t's in the future perhaps?).

Do that by making any expr2t{c,} actually a reference to the base_type
or base_container. This seems to work. Unfortunately, pointer_type2tc
works in the same way as not2tc, you can't upcast a type2tc to a
pointer_type2tc, only construct a new one. Too bad.

This patch replicates the type macro scenario for types, because
apparently, due to something I've done, we can only have dynamic casts
of the type objects once they've been declared with a complete type.
(This seemed to work in the past though. Never mind).