Fix type solving in part of type2tc.
Basically, the compiler doesn't seem to be able to work out that what
the base container returns from operator-> in the changed expression is
of the base_type class. This might be that it knows it, but because
we're not explicitly naming it as base_class, it doesn't match it.

Getting the pointer explicitly so that it knows its type works.