Fix double pointer type setup situation.
What I wanted here was a double pointer to the string type. What I
actually got was a pointer to the string type, that was then copy
constructed into another pointer type, because it was passed a copy of
itself.

In theory this could be disabled by removing the explicit copy
constructor, and the other downcast constructor should be disabled too
by the template situation. Not working for some unknown reason.