Fix how slicing occurs.
In the past (pre irep2) we took the identifier of the thing we were
slicing, and kept that in a hash set. Fine.

However with irep2, I moved to the name record thing, on account of how
it's a better storage for names. But, I missed the fact that that object
was only storing the l1 part of the name (which is what's needed in the
level2 identity map).

What that means is that once one l1 name makes its way into the depends
set, all values of that variable get kept by the slicer. Which is bad.
So, store an expr2tc, normalising the type so that slightly different
pointer types of a variable won't make a difference.