Optimise slicing a little.
Storing all variable names in a hash set is bad: instead, store all
names for the period that they're alive. At times before they've been
defined, they won't be accessed by other assignments, so remove them
from the depends set (which reduces the number of comparisons that have
to occur).

This fix, in combination with the previous commit, knock the tacas
verification time down from 179ks to 163ks, which is very reasonable.