Stop object numbering from using vectors.
Instead, for looking up the number of the contained number, use a hash
map with the integer value used as the hash key. Most of the time this
is going to result in constant-time lookup anyway, as the integers are
going to increase linearly.

The point of this that we can erase items from both of these containers
without having to do any data juggling, and can just have a
monotonically increasing object number.