Start remunging the value set obj numbering thing
I ran massif on one of the TACAS tests, and it suggested that a /large/
amount of memory is stashed by a call to get_original_name in
goto_symex_statet::assignment, which I figure is because it renames the
rhs of an assignment to l1, then gets value_sett to store it somewhere.

That somewhere is value_sett::object_numbering, which never releases
anything it stores. Which is bad. So, I'm going to keep some separate
reference counting indicating how many times an expression in there is
currently used, and remove it when we're done. Value set names are
already deleted once we pop a stack frame in symex, so we should be able
to piggyback on that.

(This is why the previous few commits exist, making numbered objects
 eraseable).