Make a fundemental type assertion.
It turns out that when assign-constructing things to date, we've been
actually calling some variety of constructor instead of the assignment
operator. Which has led to no assertions being made regarding types.
Which has just led to me debugging something for half an hour that
should have immediately trapped. Which is annoying.