Generate an irep2 C implicit typecast impl.
It seems that the actual code and logic in that module is a) useful b)
relatively concise enough, so there's no point trying to rewrite it,
just duplicate it instead and move all of the types over to irep2.

One failure is that we don't represent qualifiers at any point in irep2,
therefore that code is removed. Won't make a difference to anything at
symex time.