Fix a linking ecentricity.
What I think is happening here is that there are two definitions of the
default constructor for z3::context, the one from an external z3++.h
that's included by metasmt, and the one that's internal to ESBMC. Both
of them live in object files, and the linker is picking one? Wheras what
we actually want is one for our own Z3 backend, and one for metasmt,
which relies on the default Z3 behaviour.

Original commit message: "Inject some extra hurrr into ESBMC."