Switch metaSMT ast storage.
So: I've taken a look through the documentation for boost::proto, on
which metaSMT is based, and it's hopeless. The entirety of the AST is
stored in a type, and the transmission of data between functions is in
the template parameter. On which I Have Opinions.

boost::proto mandates that the "Context" (in this case Z3_Backend) has
some operator() overloads that take an AST identifier (as a type),
and some result_type operands, and which returns a result_type.  The
idea being that you evaluate the AST node by the identifier, using the
operands, and produce a result. So that's the bones of the evaluation of
the expressions, evaluating into the type result_type.

This is important to us because we can't store normal expressions (which
are _types_, for crying out loud!) across the SMT conversion interface,
and we absolutely have to store it as some kind of known data type. So
we're going to end up manually evaluating and building expressions using
the operator() interface as above, constructing a series of
result_type's, which will represent the AST node itself. By coincidence,
result_type is a z3::expr.

It isn't pretty, but it's the only way to work with this behemoth.