Remunge symbol creation.
Turns out that by using new_variable, new_array and so forth, we're
creating proto'ified variables, and all expressions containing those
variables get tainted with the operator overloading. Which means we just
can't use the bloody things.

So instead copy their implementation by picking out the underlying data
class and operate on that instead. The operator overloading stuff in
Z3_Backend copes with this just fine, and we can munge the symbol table
gunk to make that work too.