Commit what made bvint-creation work.
So it looks like the correct approach is going to be interacting
_directly_ with the solver backend, through the metaSMT defined data
interface.

This is because /whenever/ I try to pump anything at /any/ method to do
with ctx, it gets operator overloaded into some kind of boost::proto
mush, everything expects a proto type expression, and the world ends.

So instead, observe that all of the required API is implemented by the
backends, with operator()'s methods (which do the actually work)
selected between via means of the type of one of the parameters. It
looks like if we ditch the DirectSolver_Context wrapper, then we can
start calling those methods almost directly.