Handle bools <=> bits better.
The problem being that under strict QF_AUFBV you can't have arrays of
bools, only arrays of bits, so we have to switch between the two. To fix
this, inject special cases (we already had a few), but make them decide
based on a flag passed down from the subclass.