Attempt to handle array equalities.
Basically they only happen when we're _assigning_ something to a symbol.
So attempt to replicate that by storing the AST into the symbol table.

This is highly liable to break in the future. When that happens, we need
to be looking at other solvers that metaSMT::z3, to see whether or not
they support array equalities. We can wrangle Z3 to do it via the back
door, if other solvers can't, we're screwed.