Implement l_get.
And by implement, I mean don't implement, always return false. The
result is that no paths through the program are true, so no further
counterexample is printed.

Right now I'm more interested in verifying that we're doing the right
thing than getting the counterexample out of these solvers. It seems
that the read_value method will allow us to extract model data out of
the context, but munging that situation may cause a days worth of brain
damage.