Start looking at handling arrays of arrays.
Right now they work because Z3 accepts them. What we should be doing is
congealing them into one array, whereby the array index is a combination
of all the indexes at each dimension of the array. Unfortunate and
unpleasent perhaps, but not that much of a problem in theory.

This will make what are hopefully the remaining metasmt tests work.