Not all with chains are to be messed with.
For one thing, they might be tuple with's, which don't have integer
indexes. Also, a series of withs might just be the result of constant
propagation, as opposed to updating an array within an array.

So be more specific: only fiddle with a with if it's withing an array of
an array.