Open the possibility to bitblast arrays.
Contary to what MetaSMT claims to do, it doesn't actually support arrays
on things that doesn't support it on an underlying level. It /does/
have facilities for bitblasting bitvectors into propositional logic, but
not arrays.

So, we're going to have to do this ourselves (hey, CBMC does it). I
thought about putting it into MetaSMT, and logically, that would make
sense, but it'd be a little too brain damaging right now. In the long
term I think moving things there would make sense; and maybe we could
contribute it back.

But anyway; a bit-blasting we will go.