Fix array dimensional foo once and for all.
Basically: the store stuff was all rubbish, I hadn't grocked that the
with chain occurred in the update_value field of the expr, not the
source_value. Never mind.

Also, we can't convert the update_value part of a store at all on
account of type failery, so instead special case it through directly to
the switch table.