Make fix_array_index deal in terms of expressions.
Which is going to be much more flexible shortly.

Some fallout: all the tuple munging things have to change (perhaps they
need to end up being expression based too?) and as a result there's a
gap in the array handling where we have an ast based index, but no expr
index. Fix that later.