Fresh arrays are also symbols.
This is supposed to be fixing the problem of array equalities: we allow
them when we're just assigning to a symbol, but otherwise not. So, some
things of array sort have to have names.

Note that we're putting it in the symbol table; necessary because we
might look it up in the future. However, we also require an id number
for metasmt... well, bah, just use zero for the moment. Sue me.