Faff around with metasmt array_of's.
So we have a problem: in the past I've been making infinite sized arrays
(for modelling ESBMC things) work by, when we encounter a new (constant)
index for it, asserting that the initial value of the array at that
field has the initializer value. This works better than an infinite
number of stores.

That doesn't work with the bitblasted arrays I've just been writing,
because they treat accesses to unaccessed indexes as being a free value.

So, when we get an array_of, override the usual API and implement it as
either a default value in unbounded arrays, or in bounded ones just
store it at every element. This means that the array_of semantics just
work. We should probably implement it for Z3 too, because that actually
has a function for this.

What this means, is that for every solver, either the initial algorithm
(above) where we make retroactive assertions has to hold, or the
array_of handler has to be manually implemented to work.