Fix some peculiar array behaviour
See 01_cbmc_Unbounded_Array1 -- when we select an unassigned index from
an array, we have to ensure its value is consistent for future reads.
Which IMO is horrific.

Implement this by storing the results of selects :(. Massively
inefficient.

Lets look at the performance of sat solvers with and without this.