Combine ite data into with data.
Seeing how we're uniquely numbering array updates, we need them to all
be in one place, which is invalid if ite's are an array update stored
elsewhere.

So, change the array update representation to be either an ite or a
with. Relatively simple. Can't union the idx because it's not a POD.