Switch back to using literalt's for things.
I've now been enlightened as to the design decisions behind all the
literal gumpf; it's properly a wrapper around minisat literal variables,
and contains all kinds of utility functions to cope with how minisat
deals with those variable numbers. So it makes entire sense to flood it
everywhere in a SAT solver.

I still thing the rewrite situation is correct: previously we had the
superclass of all conversions being prop_convt, which contained all
kinds of SAT solving gunk, which IMO is a backwards design. (Because
we've shoehorned SMT solvers into something not designed for them).

A better design is one where the higher level classes convert
expressions into simpler representations, and we can fit in different
classes of solvers at lower points in the derived chain. So minisat
would be something like:

  smt_convt <- tuple_convt <- array_convt <- bv_convt <- prop_convt <-
  minisat_convt

But I digress.