Special case array equalities.
Seeing how we're playing silly buggers with this in the bitblasting
array representation (by changing AST data, the array num, depending on
an operation upon it) we need some kind of hook.

Specifically the problem is that, in an equality, the new array symbol
gets converted, put into the AST cache, then when it's assigned the
underlying solver updates its symbol table to make that equality
effective. But, the previous, un-updated value is going to be pulled out
of AST cache. Which leads to some conflicting behaviour.

So, make it the underlying solvers problem.