Ensure some type safety in tuple operations.
That is: if we have some expression passed in, then it might be some
horrible constant propagated somethingorother, or a constant struct, or
whatever. What we want it to be is a tuple symbol that we can
manipulate. So, hand that off to force_expr_to_tuple_sym, which converts
the expression and returns the appropriate symbol.