Pump array code into array_convt file.
This involves massacring some things. First, there are three new methods
at the top of array_convt that the instigating subclass must implement.
Secondly, is_unbounded_array becomes a macro-ish.

Finally, there's the problem of getting the range bitvector width:
that's going to have to be another virtual function for now, that things
like minisat override. Long term, all this width data should be stored
in smt_sort itself, because it's way to useful to tie up in
solver-dependant structs.