Rethink approach to the 'get' api.
Basically we sould decompose it to a series of "get this" and "get that"
at the smt_convt level, and have things implement the funcs introduced
here.

In the meantime, it's necessary for these to exist, and be virtual in
the base class, so that bitblast_convt can call through to the
array_convt sister class for array_get.