Massacre smt-api code into bitblast_convt.
And make it build, and interract correctly with minisat_convt.

Basically, there are quite a few hacks here, which are congealing around
smt_convt. This is good, because it means the api is going to get a lot
more defined, and all of the problematic cases are going to end up going
via virtual function pointers.