Fix some infinite recursion.
The problem being that AND calls land calls AND.

Fix this by not having to specify smt_convt, thus getting the overridden
land from cnf_cont. Do this by accessing via the type prop_convt, which
forces a lookup from the vtable.

This didn't work previously because we define our own overloaded land;
therefore when we call land, it expects it to be that one. I don't know
why this is the case, it seems silly, but there we are.