Start moving away from multi-inherit. to templates
So: clearly the overhead of having sister classes everywhere is going to
be absolutely rubbish, and lead to a bunch of useless declarations in
smt_convt that doesn't give any type assurance, because we can randomly
abort at any point because someone didn't implement a virtual function.
This is bad.

So instead do something inspired by metasmt: compose classes via
templates to create a chain of classes implementing the requirements of
their subclasses, templated. Unpleasent; ugly; but probably effective.