Speculate on some fixes to scanner.l
gotident doesn't appear to have an implementation. To get around this,
just print out the value of the identifier we're dealing with. I figure
this is probably what's supposed to happen here; who can say whether
it's right. I'll find out later.

prtprag doesn't have a body, and no suggestion as to what it does. So,
comment it out. Probably has something to do with pragmas.

The other two allocations are necessary because there seems to be some
confusion as to whether the bbuf field of struct includ is in fact a
static buffer, or a heap allocated hunk. Resolve this confusion.