Fix some kind of linking fail with clang.
When these functions are marked as inline, they don't get compiled into
the output. And as a result, some other lump of code that doesn't
actually inline them generates a reference to them, which is undefined,
which fails.

Pretty damn certain this is a compiler bug, but it's clang 3.0 here, so
it's quite likely to have been fixed.