Virtualize expr2t/type2t destructor.
So. Only clang complained about this not being present. In theory, the
appropriate destructor was never called, and we never freed sub-exprs.
Which seems wrong, because I'm pretty sure that this worked in the past.

This might lead to some kind of performance hit; on the other hand, it
should also lead to some kind of memory saving too.

More concerning is that gcc must have been doing something crazy to make
this work in the past.