Fix a problem with EXTRACXXFLAGS propagating
Because all the make variables enter the program environment when you
run a program in make, if you have two directories, one with
STRICTCOMPILE set, and one inside it that does not, then you get both
sets of CXXFLAGS because one inherits from the other.

To get around this, unset CFLAGS and CXXFLAGS whenever we go down a
level in a makefile. Only preserve things that were in EXTRAC{,XX}FLAGS
and don't put anything in there ourselves.