Define away __asm on macs.
It seems that there's some scenario whereby an alias qualifier gets
attached to a bunch of symbols, indicating that they alias other
existing functions. Which is fine, except that the way this is done
is by emitting strings to the assembler, I can't work out how to turn
it off, and the parser chokes on it. So, define it away instead of
attempting to handle it gracefully.