Don't fortify sources on macs.
By default, it seems, macs enable _FORTIFY_SOURCE, which turns a
bunch of string manipulation functions into macros that attempt to
correctly do bounds checking before invoking the function call.

However within ESBMC, that leads to a bunch of function names evaluating
to '0', which is poor when you're trying to either use them or define
their body. So, define this away too.