Fix the C library -> object file conversion.
The purpose of this code is to convert a binary file into an object
file with a symbol that we can use to read it in-process. That means
pumping it through 'od', mashing the output, and compiling it.
However, it turns out that the format string chosen before ('d1')
prints negative signs ('-') occasionally, as it's printing signed
characters. That led to some format muckups, which then exploded and
led to the C library not coming out as what went in.

So fix that by printing characters as unsigned chars.