Fix a spree of compiler warnings and errors.
These are all clang based, and range from objecting to unused
variables to the fact that 'uint' isn't really a type, and I don't
know where that's coming from on gcc.

The call to mk_smt_symbol in smt_tuple.cpp is still required for
some kind of legacy reason I can't remember to do with the symbol
table. If you remove it, things break randomly, I can't remember why.

The const cast on array_assign_symbol is because the const situation
for smt_ast's is entirely broken and aimless.