Discard preprocessor line-number directives.
CPP does not like these for some reason, I don't know why. The
side-effect of this is that if we preprocess a set of files all into
one, then the line-no information is going to get discarded before ESBMC
can parse it.

Which is not a bad thing on the whole, because it means you can just
look through the file we're working on, rather than having to reverse
engineer where in the parsed file and where in the included files are
what.