Avoid repeatedly checking the same assertions.
That is, if we assert something at the start of a reachability tree,
currently we'd end up verifying it in /every single interleaving/. Which
is somewhat inefficient.

So instead: once we verify an interleaving and start backtracking, wipe
out all the previously checked assertions once we settle on another
thread to run. They've been checked: they needn't be handled again. We
can't unfortunately either cache or eliminate the assertion instructions
all in one swoop, because symex_target_equationt's are duplicated when
we generate a new execution_statet. Which sucks, but there we are.

Don't do this for --schedule: they way it works already handles this
problem.