Re-install state hashing.
This is a pigs ear of a patch because I couldn't get this to compile in
some middle steps. Basically, this moves the crypto_hash api to use the
'ingest' method to accept data, and has irep2 types send data to it in
the same way as crcs.

Then, we remove all of the old hashing stuff from execution_statet, and
have it use the new expression hasher. Otherwise, the algorithm is
almost entirely the same, aside from some null guards.