Start importing code from irep2_memmodel
Code related to fixing the frontend / pointer tracking to be less
rubbish. This particular patch ensures that the size of memory objects
is measured accurately.

That immediately leads to all array bounds checks failing, because
they're busy storing array indexes in the 'pointer offset' data field.
Quite a lot of breakage is going to occur as the result of these
patches; too bad.