Import value-set/deref code from irep2_memmodel
This fixes various things, such as multi-dimensional array access (i.e.,
address it, rather than representing it in SMT) and the assertions when
dereferencing pointers to arrays, ensuring that the correct units are
used (i.e. bytes not indexes).

As I recall, this leads to many more nondeterministically offset byte
extracts and updates. We shall see.