Nurk the 'namespace' shuttling in value_sett.
It's passed around between almost every single method, but only used by
very few. It's also constant between model checking contexts. So, make
it a member of value_sett, and juggle parameters appropriately.

I was going to remove it, but it seems it's necessary for calls to
from_expr in value_sett::output, and some calls to base_type_eq, in
particular is_subclass_of is fundementally required.