Make educated guess at pointer alignment updates
i.e., how to update and merge some of the alignments. This introduces
the idea of the natural alignment, i.e. how to reduce a 98 byte offset
into a structure into an alignment that makes sense. I haven't really
thought it through yet though.

This is all going to burst into flames when running into the problem
that none of the structs we allocate nowdays have padding.