Kill some dead ansi-c code.
It would appear that it's supposed to be dedicated to coping with taking
the size of an array dimension / slice. However a) it's commented out,
b) we don't know why, c) we can deal with that when we run into it.

Or, possibly, a better approach, of just flattening array calculations
in the frontend.