Refactor collection of scalar steps.
The problem first encountered is that the 'base', i.e. what we're
dereferencing, isn't immediately available to decompose_top_scalar_expr.
It's also difficult to determine, because it requires looking at indexes
to work out whether they're dereferences too. So, in general, is a bad
plan.

So instead: we have symex_dereference construct the set of scalar steps
required, and at the same time have it flush out any if2t's that occur
in the dereference expression. This is much cleaner.