Join two competing dereference_rec implementations
It looks like the one in goto_program_dereference was existing only to
provide assertion callbacks, and to accurately represent the guards of
when the expressions are /actually/ evaluated. (i.e.: short circuiting,
tertiary operator, etc). Wheras the goto_symex one is just for getting
expressions.

This is code replication, and is only going to get worse, so merge it
into the dereferencet class. This also means I now know what was
breaking my previous attempts to get rid of the static analysis
(evaluation guards), so can get on with that...