Comment out some pointer checks from io.c
These appear to eventually get discarded from the program, but they
exist here so that they a) evaluate to a struct and get discarded, b )
trigger all the pointer checking. Which is good. But derefing to a
struct isn't a thing right now.

Will need to handle this in the future.