Fix handling string types.
The to_array() method of a constant string returns a temporary expr2tc,
which evaporates the moment the line is finished. Which means the
reference to its type goes away; this is bad.

So, we only actually want the subtype, so extract the subtype directly
and store a reference to it.