Correct calculation of final dereference type.
We want to dereference either a scalar, or through a list of scalar
steps to reach a scalar. Thus, set the dereference type to that final
scalar type.

This is necessary because invalid symbols, ite operations, and so forth
all rely on this type being the final access type, rather than an
intermediate nonscalar.