Add method turning a nonscalar expr into a scalar.
It does this by repeatedly accessing the first element in the array or
struct, or whatever. The reasoning here being that that element or field
has to exist (by the very demands of the C spec). And that element
should always be in range / accessible if the pointer is correct.

The purpose of this code is to act as a vehicle for a dereference of a
non-scalar expression, which is only existing to check that the pointer
being used is valid. For example:

  struct foo *bar = NULL;
  *bar;

Is perfectly legitimate C: the outcome of dereferncing bar is discarded,
but as it's dereferenced it should trigger a null pointer dereference
failure. Currently this isn't possible because you can't access a
nonscalar in a dereference, but this should allow it in circumstances
where we know that the result isn't interpreted.