Remunge fabricate_scalar_access to work.
We don't actually have a source object: that only happens after we look
at the value set. What we were passing down was something of pointer
type. Which was silly.

So instead, discard the idea of operating on an object, and stuff empty
exprs into member and index operations. We'll have to document this in
the future.