Fix some pointer arithmetic.
This is potentially the cause of quite a lot of faults: the type size
multiplication of one side of pointer arithmetic was being dropped.
Which leads to some seriously incorrect code as a result.

In the previous revision, digital_filter_01 triggers this fault, as it
gets an addition of a pointer to '1'; but at a higher level this is
divided by 4. The result is the '1' being dropped by the divide.