Cope with byte_updating two dimensional arrays.
By constructing the necessary selects and stores around it. This is only
intended as a sticking plaster: the code in irep2_smtlib is much much
better at getting this done, as it's going to achieve flattening all
these multidimensional arrays down to a single dimension array. Which
will be good.