'Fix' a depth of a regression test.
Basically, this test has only ever not taken several minutes because the
depth is set in such a way that it never hits any of the claims that the
solver needs to test. If it's a large depth, it takes a long time, and
finds a bug.

While this is triggered by the memory model changes I've made, it's not
caused by them. So fix this by shortening the depth so that it does what
it did in the past. I will hand in my degree certificate at the door.