Switch outcome of a pointer regression test.
So, previously this had a verification failure because of an assertion
that pointer relations are only defined on the same object. So there's
an assertion baked into this program to ensure that the relation in
st_compact is the same object.

However, it's not actually a violation -- that field is always
statically null, so relating it to zero isn't a problem. Not certain
what I changed to make this work, but the previous assertion failure was
incorrect, IMO.

CBMC doesn't report a failure either.