Accumulate some missed offsets into pointer derefs
i.e.: if we have the code in 01_cbmc_Pointer10, of:

  struct foo {
    int a;
    int array[10];
  };

and take the address of the 4rd element of array, then what the pointer
tracking code coughs up as the 'what' argument to build_reference_to, is
an expression ending at 'array', rather than a particular struct. So
there are 4 additional bytes of offset that are hidden.

Try to get hold of these via the use of compute_pointer_offset.