Add first attempt at handling dynamic struct offs.
Current algorithm is to go through each field in the struct, and encode
all the selects that might be made against it, according to the given
alignment and access size. The idea being that we don't enumerate every
offset, we enumerate all the fields and work out the different accesses
that can be made against them.

This, in combination with not considering unaligned / incorrect
accesses, should decrease the behaviours present in previous approaches.
Hopefully.

Currently it computes the set of things to encode, but doesn't combine
them together into one value.