Fix an off by one.
Consider: accessing the end of a 32 bit integer as a character. In that
case, int_offset is 3, m_offs is (say) 0, (int_offset - m_offs) is 3,
access_size is 1, m_size is 4. So the second clause in the modified if
simplifies to 4 < 4, which isn't true, even though the access is valid.