Start working on struct dereferences.
That is, allowing us to dereference to a struct type. There is actually
code (31_cdaudio) that is making use of this feature, sadly.

This works by not going down any of the existing code paths, and
attempting to throw out the dereference at every opportunity as being
misaligned or incompatible.