Stop using the pointer analysis for assertions.
This doesn't run the analysis, and attempts to use claims encoded at
runtime instead. Several more regression tests fail as a result, but
they seem to be for valid reasons -- for example a double free in
01_cbmc_Free6. On the whole this seems to be a beneficial change, and
makes the 23_picosat-846_0* tests work again. We might even be able to
delete a lot of complexity from all this code...

The patch itself is a pigs ear, as I'm yet to delete some of this code,
I'm committing half way through as I keep on knackering things. The
patch:
 * Disables the pointer static analysis.
 * Shoehorns a dereference in 'free' mode into symex_free (previously
   this was handled by assertions encoded by the pointer analysis).
 * Encode claims reported at symex time dereference.
 * Moves calls to replace_dynamic_allocation to after dereferencing --
   these were previously massaged during the pointer analysis, but need
   to happen after dereferencing.
 * Moves access-to-freed-dyn-object check to being encoded in
   valid_check.