Attempt to wrap rewritten ptr types in casts.
If we dereference something from a uint32 pointer, and it ends up being
a uint8 type, then we still need to do some pointer arith on the fact
that the type system thinks it's a uint32 pointer. Thus this cast.

Unfortunately, it may explode due to the wretched type renaming
situation.