Remove an array lower bound check.
This was breaking 03_nec37, which had a (legitimate) negative number as
an index. This because if you interpret large unsigned indexes as
signed, you get negatives.

Clearly this is broken, and we should just check that the index is in
the range of zero to the size of the array (by removing the 'lower
bound' check). However, 01_cbmc_Malloc28 now breaks, but for interesting
overflow reasons.

Either way, this patch is progress.