Attempt to cope with unions.
These are always going to be caught by either scalar step lists or
structure refs. So this will probably catch a bunch of cases?

The problem is going to be when we have a pointer into a union, and need
to work out what field in the union it points at. Which will explode.
Possibly it's better to turn all of these things into byte models...