Widen num of bases accepted by struct ref derefs.
See the C++ test esbmc-stack/stack_class. This has a class within an
array within a class. We intiailly failed to recurse into the array when
the offset was zero; instead of rejecting this case, recurse into it to
check.

Success is stil the case where the offset is zero and the struct type
accessed is compatible.