Cope with incopmlete array type conversions :/
These crop up when we have the following:

  extern char const   *pci_power_names[] ;

Which essentially declares an array of unknown size that's stored
externally. Not much we can do about the size of it, so just assume that
it's infinite.

Now, thish should do some type checking to verify that only arrays
declared in such a way get this infinitely sized type, however I can't
get that working right now.