Handle code types with copmlicated returns.
The problem that i'm encountering is in TACAS14, where we have:

  extern char const * const *
  cx2341x_ctrl_get_menu(struct cx2341x_mpeg_params const *, u32);

Where the problem is the return type. The last const there is ending up
being applied to the function decl itself, not the pointer that's being
returned. The pointer record itself is being applied to the 'code' type
(in the subtype field), and is ending up being ignored.

So, juggle merge_types to store types in the subtype field of code
records.