Fix a sign error when testing for negative ints.
We were using this large number to check to see whether or not a number
was in the negative range of a 64 bit value. But, as we were storing it
in an int64, we were creating the lowest value possible, making the
following comparison always true. Which led to problems.

This turned out to be masked, repeatedly, by most values being reduced
to 32 bits implicitly by 32 bit arithmetic that was simplified. This
then led to explosions in TACAS14 on irep2_memmodel2, which seems to
trigger more code behaviours.